+++ src/main/java/org/cicirello/search/evo/FitnessBiasFunction.java
@@ -22,7 +22,8 @@ package org.cicirello.search.evo;
 
 /**
  * This functional interface is used to provide a bias function to the 
- * {@link BiasedFitnessProportionateSelection} operator.
+ * {@link BiasedFitnessProportionateSelection} operator as well as the
+ * {@link BiasedStochasticUniversalSampling} operator.
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>