+++ src/test/java/org/cicirello/search/evo/SelectionOperatorTests.java
@@ -219,6 +219,104 @@ public class SelectionOperatorTests {
 		}
 	}
 	
+	@Test
+	public void testBiasedSUS() {
+		BiasedStochasticUniversalSampling selection = new BiasedStochasticUniversalSampling(x -> 11*x);
+		validateIndexes_Double(selection);
+		validateIndexes_Integer(selection);
+		StochasticUniversalSampling selection2 = selection.split();
+		validateIndexes_Double(selection2);
+		validateIndexes_Integer(selection2);
+		
+		validateHigherFitnessSelectedMoreOften_Double(selection);
+		validateHigherFitnessSelectedMoreOften_Integer(selection);
+		
+		double[] weights = selection.computeWeightRunningSum(new PopFitVectorDoubleSimple(5));
+		double[] expected = {11, 33, 66, 110, 165};
+		assertEquals(5, weights.length);
+		for (int i = 0; i < weights.length; i++) {
+			assertEquals(expected[i], weights[i], 1E-10);
+		}
+		
+		weights = selection.computeWeightRunningSum(new PopFitVectorIntegerSimple(5));
+		assertEquals(5, weights.length);
+		for (int i = 0; i < weights.length; i++) {
+			assertEquals(expected[i], weights[i], 1E-10);
+		}
+		
+		PopFitVectorDouble pf = new PopFitVectorDouble(16);
+		int[] selected = new int[16];
+		selection.select(pf, selected);
+		int[] expectedMin = new int[16];
+		int[] expectedMax = new int[16];
+		int[] counts = new int[16];
+		double totalFitness = 0;
+		for (int i = 0; i < 16; i++) {
+			totalFitness = totalFitness + pf.getFitness(i);
+			counts[selected[i]]++;
+		}
+		for (int i = 0; i < 16; i++) {
+			expectedMin[i] = (int)(16 * pf.getFitness(i) / totalFitness);
+			expectedMax[i] = (int)Math.ceil(16 * pf.getFitness(i) / totalFitness);
+			assertTrue("i:"+i+" count:"+counts[i]+" min:"+expectedMin[i], counts[i] >= expectedMin[i]);
+			assertTrue("i:"+i+" count:"+counts[i]+" max:"+expectedMax[i], counts[i] <= expectedMax[i]);
+		}
+		
+		PopFitVectorInteger pf2 = new PopFitVectorInteger(16);
+		selected = new int[16];
+		selection.select(pf2, selected);
+		expectedMin = new int[16];
+		expectedMax = new int[16];
+		counts = new int[16];
+		totalFitness = 0;
+		for (int i = 0; i < 16; i++) {
+			totalFitness = totalFitness + pf2.getFitness(i);
+			counts[selected[i]]++;
+		}
+		for (int i = 0; i < 16; i++) {
+			expectedMin[i] = (int)(16 * pf2.getFitness(i) / totalFitness);
+			expectedMax[i] = (int)Math.ceil(16 * pf2.getFitness(i) / totalFitness);
+			assertTrue("i:"+i+" count:"+counts[i]+" min:"+expectedMin[i], counts[i] >= expectedMin[i]);
+			assertTrue("i:"+i+" count:"+counts[i]+" max:"+expectedMax[i], counts[i] <= expectedMax[i]);
+		}
+		
+		PopFitVectorIntegerSimple pf3 = new PopFitVectorIntegerSimple(16);
+		selected = new int[16];
+		selection.select(pf3, selected);
+		expectedMin = new int[16];
+		expectedMax = new int[16];
+		counts = new int[16];
+		totalFitness = 0;
+		for (int i = 0; i < 16; i++) {
+			totalFitness = totalFitness + pf3.getFitness(i);
+			counts[selected[i]]++;
+		}
+		for (int i = 0; i < 16; i++) {
+			expectedMin[i] = (int)(16 * pf3.getFitness(i) / totalFitness);
+			expectedMax[i] = (int)Math.ceil(16 * pf3.getFitness(i) / totalFitness);
+			assertTrue("i:"+i+" count:"+counts[i]+" min:"+expectedMin[i], counts[i] >= expectedMin[i]);
+			assertTrue("i:"+i+" count:"+counts[i]+" max:"+expectedMax[i], counts[i] <= expectedMax[i]);
+		}
+		
+		PopFitVectorDoubleSimple pf4 = new PopFitVectorDoubleSimple(16);
+		selected = new int[16];
+		selection.select(pf4, selected);
+		expectedMin = new int[16];
+		expectedMax = new int[16];
+		counts = new int[16];
+		totalFitness = 0;
+		for (int i = 0; i < 16; i++) {
+			totalFitness = totalFitness + pf4.getFitness(i);
+			counts[selected[i]]++;
+		}
+		for (int i = 0; i < 16; i++) {
+			expectedMin[i] = (int)(16 * pf4.getFitness(i) / totalFitness);
+			expectedMax[i] = (int)Math.ceil(16 * pf4.getFitness(i) / totalFitness);
+			assertTrue("i:"+i+" count:"+counts[i]+" min:"+expectedMin[i], counts[i] >= expectedMin[i]);
+			assertTrue("i:"+i+" count:"+counts[i]+" max:"+expectedMax[i], counts[i] <= expectedMax[i]);
+		}
+	}
+	
 	private void validateHigherFitnessSelectedMoreOften_Double(SelectionOperator selection) {
 		// This part of the test attempts to confirm that greater weight is placed on
 		// higher fitness population members. A sporadic failure is not necessarily a